@@ -4,6 +4,16 @@
   "private": true,
   "author": "robbyra@gmail.com",
   "dependencies": {
+    "@grafana/faro-core": "^1.0.0",
+    "@grafana/faro-react": "^1.0.0",
+    "@grafana/faro-web-sdk": "^1.0.0",
+    "@grafana/faro-web-tracing": "^1.0.0",
+    "@opentelemetry/api": "^1.4.0",
+    "@opentelemetry/exporter-trace-otlp-http": "^0.35.1",
+    "@opentelemetry/resources": "^1.9.1",
+    "@opentelemetry/sdk-trace-base": "^1.9.1",
+    "@opentelemetry/sdk-trace-web": "^1.9.1",
+    "@opentelemetry/semantic-conventions": "^1.9.1",
     "@reduxjs/toolkit": "^1.6.0",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
@@ -12,17 +22,19 @@
     "@types/jest": "^26.0.15",
     "@types/node": "^12.0.0",
     "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
+    "@types/react-dom": "^18.0.11",
     "@types/react-redux": "^7.1.16",
     "@types/styled-components": "^5.1.10",
     "axios": "^0.21.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.5",
+    "react-router-dom": "^6.8.1",
     "react-scripts": "^5.0.1",
     "redux": "^4.1.0",
     "redux-saga": "^1.1.3",
     "styled-components": "^5.3.0",
+    "styled-reset": "^4.4.5",
     "typescript": "4.1.3",
     "web-vitals": "^1.0.1"
   },
