@@ -1,2 +1,5 @@
 export const API_KEY = process.env.REACT_APP_WEATHER_API_KEY;
-export const API = process.env.REACT_APP_WEATHER_API;
+export const WEATHER_API = process.env.REACT_APP_WEATHER_API;
+export const FARO_URL = process.env.REACT_APP_FARO_URL;
+export const ITEM_API = process.env.REACT_APP_ITEM_API;
+export const OTEL_COLLECTOR_URL = process.env.REACT_APP_OTEL_COLLECTOR_URL;
